Moved menu to Verification & Heaps of Additions
Moved the Menu to only be called from OAuthVerification to allow for a
better experience overall.

Only when a user is verified/logged in, will they see all the menu
options for application control.

Also added fully functional Player Controls, You can control Volume,
Next/Previous & anything inside the PlayerCtrl scope.

The volume control is a bit messy, But i wasn’t able to have an easy
approach with adjustVolume in the PlayerCtrl scope.

Added a DevTools shortcut in the Help Menu, Easily removed if needed
(being in the dev branch, i thought it would be an easy way to access
the DevTools if you didn’t know the keyboard shortcut)