[Feature] Added adjustable playback rate.
This is especially useful for those of us who use Soundcloud for listening to podcasts. Any respectible podcast app allows you to change the playback rate (typically up to 3.5x).

The implementation is almost identical to the volume slider.

Playback rate is saved to localStorage same as volume.

Possible improvements (todo):
- Add numbers to the slider such that the user can see precisely what speed they are playing at.
- Tweak styles to fit in with the UI better (currently it uses the same styles as the volume slider).